Moved GQL schemas to separate files.
