+++ src/schema/typeDefs/animal.graphql
@@ -1,0 +1,81 @@
+extend type Query {
+  """
+    Lookup an animal.
+  
+    Examples:
+  
+    animal(id: 1)
+  """
+  animal(
+    "Animal id in database"
+    id: Int!) : Animal
+
+  """
+    Get all animals.
+  
+    Examples:
+  
+    animals
+  """
+  animals : [Animal]
+}
+
+"Represents an animal."
+type Animal {
+  "Animal id, for example 2"
+  id: Int!
+  "Animal name"
+  name: String
+  "Organization id"
+  organization: Int!
+  "Status"
+  status: String
+  "Image URL"
+  imageUrl: String
+  "Comments"
+  comments: String
+  "Modification time"
+  modTime: String
+
+  "Animal active registration info"
+  registration: AnimalRegistration
+  "Animal implanted microchip info"
+  microchip: AnimalMicrochip
+  "Animal details"
+  details: AnimalDetails
+}
+
+extend type Mutation {
+  "Created animal"
+  createAnimal(input: CreateAnimalInput!): Animal
+  "Updated animal"
+  updateAnimal(input: UpdateAnimalInput!): Animal
+}
+
+input CreateAnimalInput {
+  "Animal name"
+  name: String
+  "Organization id"
+  organization: Int!
+  "Status"
+  status: String
+  "Image URL"
+  image_url: String
+  "Comments"
+  comments: String
+}
+
+input UpdateAnimalInput {
+  "Animal id, for example 2"
+  id: Int!
+  "Animal name"
+  name: String!
+  "Organization id"
+  organization: Int!
+  "Status"
+  status: String!
+  "Image URL"
+  image_url: String!
+  "Comments"
+  comments: String!
+}
