+++ src/schema/typeDefs/animalDetails.graphql
@@ -1,0 +1,33 @@
+"Represents an animal details."
+type AnimalDetails {
+  "Animal id, for example 2"
+  animalId: Int!
+  """
+  Animal breed by language.
+  Examples: breed(language: "en") or just breed - will return default language ("${defaultLanguage}") translation
+  """
+  breed("Language code" language: String = "${defaultLanguage}"): String
+  """
+  Animal species by language
+  Examples: species(language: "en") or just species - will return default language ("${defaultLanguage}") translation
+  """
+  species("Language code" language: String = "${defaultLanguage}"): String
+  """
+  Animal gender by language.
+  Examples: gender(language: "en") or just gender - will return default language ("${defaultLanguage}") translation
+  """
+  gender("Language code" language: String = "${defaultLanguage}"): String
+  """
+  Animal color by language
+  Examples: color(language: "en") or just color - will return default language ("${defaultLanguage}") translation
+  """
+  color("Language code" language: String = "${defaultLanguage}"): String
+  "Animal date of birth"
+  birthDate: String
+  "Animal weight (kg)"
+  weight: Float
+  "Animal allergy"
+  allergy: String
+  "Animal food"
+  food: String
+}
