+++ src/schema/typeDefs/color.graphql
@@ -1,0 +1,18 @@
+extend type Query {
+  """
+  Get all colors.
+
+  Examples:
+
+  colors(language: "lt")
+  """
+  colors("language code" language: String!): [Gender]
+}
+
+"Represents a color."
+type Color {
+  "Color id"
+  id: Int!
+  "Color name"
+  value: String!
+}
