+++ src/schema/typeDefs/gender.graphql
@@ -1,0 +1,18 @@
+extend type Query {
+  """
+  Get all genders.
+
+  Examples:
+
+  genders(language: "lt")
+  """
+  genders("language code" language: String!): [Gender]
+}
+
+"Represents a gender."
+type Gender {
+  "Gender id"
+  id: String!
+  "Gender name"
+  value: String!
+}
