+++ src/schema/typeDefs/mutation.graphql
@@ -1,0 +1,3 @@
+type Mutation {
+  _empty: String
+}
