+++ src/schema/typeDefs/query.graphql
@@ -1,0 +1,3 @@
+type Query {
+  _empty: String
+}
