+++ src/schema/typeDefs/species.graphql
@@ -1,0 +1,18 @@
+extend type Query {
+  """
+  Get all species.
+
+  Examples:
+
+  species(language: "lt")
+  """
+  species("language code" language: String!): [Species]
+}
+
+"Represents a breed."
+type Species {
+  "Species id"
+  id: String!
+  "Species name"
+  value: String!
+}
