+++ src/schema/typeDefs/status.graphql
@@ -1,0 +1,18 @@
+extend type Query {
+  """
+  Get all statuses.
+
+  Examples:
+
+  statuses(language: "lt")
+  """
+  statuses("language code" language: String!): [Status]
+}
+
+"Represents a status."
+type Status {
+  "Status id"
+  id: String!
+  "Status name"
+  value: String!
+}
